refactor(log-viewer): move inspector empty-state copy to a leaf module
`detailSections.ts` imports the section builders' web components, so tests that
touch it must mock the whole module — which forced `LogInspector.test.ts` to
hand-mirror the four empty-state strings in its mock factory. The copy could
then drift from the real thing with the test still passing.

`detailEmptyText.ts` holds `emptyTextFor` with a type-only import, so consumers
and their tests reach the real copy without pulling in Tabulator. The mirror and
the `emptyTextFor` mock entry are gone, and its tests move to their own suite.
