+++ log-viewer/src/components/LogInspector.ts
@@ -7,7 +7,8 @@ import { customElement, property, state } from 'lit/decorators.js';
 import { type DetailSelection, type DetailSource, eventBus } from '../core/events/EventBus.js';
 import { debounce } from '../core/utility/Util.js';
 import { getSettings, updateSetting } from '../features/settings/Settings.js';
-import { buildDetailSections, emptyTextFor } from './detailSections.js';
+import { emptyTextFor } from './detailEmptyText.js';
+import { buildDetailSections } from './detailSections.js';
 import { globalStyles } from '../styles/global.styles.js';
 import type { DockPosition } from './DetailDock.js';
 import './DockLayout.js';
