+++ log-viewer/src/components/__tests__/detailSections.test.ts
@@ -19,7 +19,7 @@ jest.mock('../../features/database/components/databaseSections.js', () => ({
   },
 }));
 
-import { buildDetailSections, emptyTextFor } from '../detailSections.js';
+import { buildDetailSections } from '../detailSections.js';
 
 describe('buildDetailSections', () => {
   it('builds the shared trio for a timeline frame', async () => {
@@ -62,18 +62,5 @@ describe('buildDetailSections', () => {
     for (const source of ['timeline', 'calltree', 'analysis', 'database'] as const) {
       expect(await buildDetailSections(source, null)).toEqual([]);
     }
-  });
-});
-
-describe('emptyTextFor', () => {
-  it('names what to click, per source', () => {
-    expect(emptyTextFor('timeline')).toBe('Select a frame on the timeline to inspect it.');
-    expect(emptyTextFor('calltree')).toBe('Select a frame in the call tree to inspect it.');
-    expect(emptyTextFor('analysis')).toBe('Select a row in the analysis grid to inspect it.');
-    expect(emptyTextFor('database')).toBe('Select a SOQL, DML or SOSL row to inspect it.');
-  });
-
-  it('falls back to a generic message when no tab is active', () => {
-    expect(emptyTextFor(undefined)).toBe('Select a row to inspect it.');
   });
 });
